Refact: Rename types and split large component
